# Receiver support binaryType ‘fragments’ …