- emit message fragments array to avoid copyfull concat

Note:
binaryType used to affect before only in WHATWG API emulation
and the optional conversion to ArrayBuffer was done inside MessageEvent.
However in this change the receiver emits messages according to binaryType
so now also EventEmitter listeners added with on(‘message’, data => {})
will get the data according to binaryType.