@guymguym 2017-02-26
@lpinca Hi, would love to get your review.
See the Note I wrote above - let me know if you think this is a breaking change or not, and if you think it's a good direction.
Thanks