@lpinca 2017-02-26
Can you please run the parser benchmark and post the results before and after this change? Thanks.