@guymguym 2017-02-26
Before:
```
$ git checkout master
Switched to branch 'master'
Your branch is up-to-date with 'origin/master'.

$ node bench/parser.benchmark.js 
ping message x 1,516,912 ops/sec ±0.66% (90 runs sampled)
ping with no data x 14,433,226 ops/sec ±1.00% (88 runs sampled)
close message x 12,816,502 ops/sec ±1.34% (87 runs sampled)
masked text message (20 bytes) x 1,010,852 ops/sec ±1.13% (88 runs sampled)
binary data (125 bytes) x 1,284,809 ops/sec ±1.29% (90 runs sampled)
binary data (65535 bytes) x 186,970 ops/sec ±1.17% (87 runs sampled)
binary data (200 KiB) x 69,283 ops/sec ±1.10% (86 runs sampled)
binary data (1 MiB) x 12,406 ops/sec ±0.90% (89 runs sampled)

$ node bench/parser.benchmark.js 
ping message x 1,500,888 ops/sec ±0.68% (89 runs sampled)
ping with no data x 13,664,804 ops/sec ±1.78% (83 runs sampled)
close message x 12,325,815 ops/sec ±1.19% (85 runs sampled)
masked text message (20 bytes) x 945,751 ops/sec ±0.98% (91 runs sampled)
binary data (125 bytes) x 1,219,381 ops/sec ±1.06% (87 runs sampled)
binary data (65535 bytes) x 184,835 ops/sec ±1.39% (87 runs sampled)
binary data (200 KiB) x 69,053 ops/sec ±1.00% (86 runs sampled)
binary data (1 MiB) x 12,375 ops/sec ±0.96% (87 runs sampled)
```

After:
```
$ git checkout guymguym-receiver-fragments 
Switched to branch 'guymguym-receiver-fragments'
Your branch is up-to-date with 'origin/guymguym-receiver-fragments'.

$ node bench/parser.benchmark.js 
ping message x 1,537,659 ops/sec ±0.98% (89 runs sampled)
ping with no data x 14,070,824 ops/sec ±0.85% (87 runs sampled)
close message x 12,873,083 ops/sec ±1.19% (85 runs sampled)
masked text message (20 bytes) x 944,098 ops/sec ±0.89% (87 runs sampled)
binary data (125 bytes) x 1,404,478 ops/sec ±1.30% (85 runs sampled)
binary data (65535 bytes) x 188,281 ops/sec ±1.31% (84 runs sampled)
binary data (200 KiB) x 70,403 ops/sec ±1.05% (86 runs sampled)
binary data (1 MiB) x 12,387 ops/sec ±0.91% (87 runs sampled)

$ node bench/parser.benchmark.js 
ping message x 1,528,134 ops/sec ±0.96% (90 runs sampled)
ping with no data x 14,189,599 ops/sec ±0.87% (89 runs sampled)
close message x 12,886,198 ops/sec ±1.64% (88 runs sampled)
masked text message (20 bytes) x 1,048,379 ops/sec ±1.38% (89 runs sampled)
binary data (125 bytes) x 1,412,051 ops/sec ±0.91% (89 runs sampled)
binary data (65535 bytes) x 189,763 ops/sec ±1.09% (89 runs sampled)
binary data (200 KiB) x 70,405 ops/sec ±1.15% (86 runs sampled)
binary data (1 MiB) x 12,353 ops/sec ±0.86% (86 runs sampled)
```