@guymguym 2017-02-26
Travis failed for node v4 but I think it's transient.
I don't have the option to re-trigger without commit, do you?