@lpinca 2017-02-26
Yes, test seems to be flaky. The timeout is probably too short. I'll restart.

I'll keep this open for a while so others can weigh in. If there are no objections I'll merge this tomorrow.